01 - Intern Persistence
In this class, we explore the concept of Internal Persistence, explore the magical world of Packages, learn to check which is the best package for our project and install SQFLite in the code! Let's replay what we've learned?

Know what internal persistence is: The act of saving information on the device. In our case, as application developers, we need to understand how to manipulate files for reading and writing on mobile devices. This way, our application can communicate with a fixed database on the device and retrieve information.

Know what packages are: They are libraries created by community members. Such packages have the power to add functionality to our project! There are numerous possibilities and you can create a package and make it available too!

Install SQFLite: The package we chose to implement internal persistence in our application. SQFLite helps us create a sequential database (SQL) on our device, as well as facilitating communication between the generated file and our application in Flutter.