03 - CRUD: read files
We have already passed half the course and the rhythm is wonderful! In this class, we learned some really cool applications, let's review?

CRUD: Database manipulation operations, CRUD is an acronym coming from English: Create, Read, Update, Delete - which translated means respectively, Create, Read, Update and Delete. We learned the concepts of CRUD and how they work in the sqflite package.

Search Information: We started our implementation of CRUD by Read and saw how to implement the find() and findAll() functions that search for database information through the query() method available for Database-type objects, imported from the sqflite package.

Transform Objects: We noticed the need to work with objects of type Map<String,dynamic> and of type List<Task> when we engage with the database. Thus, so that our project can communicate with the database, we learn to transform map-type objects - coming from the database - into Task List objects - used to build the screen.