04 - CRUD: Create, Update, Delete
We are mastering CRUD now with all the methods we have created, aren't we? But to really master, we need to review what we have learned!

Insert: Essential method for adding information to the database. In the sqflite package, insert() requires two mandatory parameters: the table name [String table] and the information to be entered in map format [Map values]. It was used in the save() function of our DAO.

Implement the update method: Method to change, update, reformulate some existing information in the database. Used for tasks that have already been created, the update() method was used in the DAO save() function after checking the previous existence of the task to be added.

Use the delete method: To remove information from the database, we implement the delete() function capable of checking the existence of a task by name and removing it directly from the table. It is worth mentioning that using the name of the task as a control check is not the best of the alternatives and that the ideal would be to use a UUID.

Do the task to map conversion: Just as the need arose to transform maps into to-do lists in the past, we need to implement reverse "conversion". That is, the conversion of a task to a Map<String,dynamic> is necessary to enter information in the sqflite sequential database.