04 - Finalising
And to close with a golden key let's go to one last step: Look back and see how much you've learned.

FutureBuilder: Screen construction widget capable of waiting for asynchronous information. FutureBuilder necessarily needs two parameters: The future:, responsible for asynchronous information and the builder:, responsible for building the screen. It is worth remembering that we use the concept of snapshot as a variable capable of storing asynchronous information.

ConnectionState: We have learned that when dealing with databases (or any information that comes from outside our project) we have to be alerted to the waiting time of the information. Today it is very common to believe that, because it is extremely fast, we do not have a waiting time when receiving data. However, there is a time when the information is not complete and, in this case, we use ConnectionState (which is the current state of the snapshot) to build several screens depending on the state of the information.

Add information in the DB: In the crucial part of the implementation in our project, we add to the form the ability to save a task directly in the database, for this we remember the controlled data coming from the form and apply the save() function coming from the DAO and, in this way, we add a new task to the database.

Delete information in the DB: To remove Tasks in our database, the delete() method has been implemented within the Task Class, using a secondary action (onLongPress) in the Previously implemented ElevatedButton.